requirements.txt
streamlit>=1.28.0
chromadb>=0.4.22
pdfplumber>=0.10.3
google-generativeai>=0.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
sentence-transformers>=2.3.0
numpy>=1.26.0
pandas>=2.1.0
plotly>=5.18.0

